% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.flat.gps.R
\name{batch.flat.gps}
\alias{batch.flat.gps}
\title{Reformats raw Telonics GPS collar data}
\usage{
batch.flat.gps(iridium_csv.dir, save.file, save.dir, returnx, pattern)
}
\arguments{
\item{iridium_csv.dir}{The directory location of the Iridium csv file to reformat.}

\item{save.file}{Whether to save the output file.}

\item{save.dir}{The directory location of the output file.}

\item{returnx}{Whether to return a dataframe of the output.}

\item{pattern}{The pattern of the input csv file.}
}
\value{
If returnx=TRUE, returns a dataframe of GPS collar data in tidy format (flat file). If save.file argument=TRUE, saves the outout as a gzip.
}
\description{
A function import and reformat raw Telonics GPS collar datato csv file into a tidy format (flat file) for analysis.
}
\examples{
\dontrun{
batch.flat.gps("./data/raw/iridium_csv", save.file=TRUE, save.dir="../data/derived/telonics", returnx=FALSE, pattern="Complete")}
}
