% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dir.R
\name{create.dir}
\alias{create.dir}
\title{Creates a file directory for a new refuge I&M project}
\usage{
create.dir(dir, proj)
}
\arguments{
\item{dir}{Location to create new project folder.}

\item{proj}{The name of the project, to be used as the name of the root folder.}
}
\description{
A function to create a new file directory structure. The directory structure is based off the Alaska I&M biometric project directory template.
}
\examples{
create.dir(getwd(), "bear_survey")
}
