% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dublin.R
\name{dublin}
\alias{dublin}
\title{Creates a csv metadata file based on Dublin Core standard}
\usage{
dublin(project, name, email, office)
}
\arguments{
\item{project}{The name of the project.}

\item{name}{Your name.}

\item{email}{You email address.}

\item{office}{The name of your refuge or office.}
}
\description{
A function to create a metadata csv file
}
\examples{
dublin(project= "Haggis Survey", name="John Doe", email="john_doe@fws.gov", office="Haggis Refuge")
}
