% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.ats.R
\name{import.ats}
\alias{import.ats}
\title{Import and format ATS Globalstar GPS collar data}
\usage{
import.ats(file.in, dir.out)
}
\arguments{
\item{file.in}{The file directory path to a folder containing Iridium csv file to reformat.}

\item{dir.out}{The file directory path where to save the formated Rdata file.}
}
\value{

}
\description{
\code{import.ats()} imports ATS Globalstar data (txt) and reformats it for visualization and analyses.
}
\examples{
\dontrun{
import.ats("./data/raw/atsdata.txt", "./data/derived/atsdata.Rdata")}
}
