% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.telonics.R
\name{import.telonics}
\alias{import.telonics}
\title{Import and format Telonics Iridium GPS collar data}
\usage{
import.telonics(
  iridium_csv.dir,
  collprogsfile,
  save.file,
  save.name,
  returnx,
  pattern
)
}
\arguments{
\item{iridium_csv.dir}{The file directory path to a folder containing Iridium csv file to reformat.}

\item{collprogsfile}{The file directory path and file name of an Rdat file  containing information on the file schedules of individual GPS collars.}

\item{save.file}{If \code{TRUE}, saves an output file.}

\item{save.name}{The file directory path of the output file (if \code{save.file=TRUE}).}

\item{returnx}{If \code{TRUE}, returns a dataframe of the output into the workspace.}

\item{pattern}{The pattern of the input csv file.}
}
\value{
If \code{returnx = TRUE}, returns a dataframe of GPS collar data in tidy format (flat file). If \code{save.file = TRUE}, saves the outout.
}
\description{
\code{import.telonics()} imports and reformats raw Telonics GPS collar data from csv file into a tidy format for visualization and analysis.
}
\examples{
\dontrun{
import.telonics(iridium_csv.dir="./data/collardata",
collprogsfile = "./data/tcp.rda", save.file=TRUE,
save.name="../data/dat.Rdata", returnx=FALSE, pattern="Complete")}
}
