% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdeditor.R
\name{mdeditor}
\alias{mdeditor}
\title{Opens mdEditor to create and edit metadata}
\usage{
mdeditor()
}
\description{
A function to open mdEditor in a web browser from which to create or edit metadata.
}
\examples{
mdeditor()
}
