% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refugetools.R
\docType{package}
\name{refugetools}
\alias{refugetools}
\title{refugetools: A package that contains helpful functions for planning and implementing refuge I&M surveys.}
\description{
A collection of tools for Alaska refuge biologists. It contains the following functions: create.dir, mdeditor, dublin, import.telonics, and import.ats.
}
\section{create.dir function}{

 The create.dir function creates a project working directory.
}

\section{mdeditor function}{

 The mdeditor function opens the mdEditor application for creating and editing data and project metadata (json file format).
}

\section{dublin function}{

 The dublin function creates a metadata csv file to document the contents of the project working directory.
}

\section{import.telonics function}{

 The import.telonics function takes raw Telonics GPS collar data file (csv) and outputs a tidy dataframe (flat file).
}

