% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dir.R
\name{create.dir}
\alias{create.dir}
\title{Creates a File Directory for a New Refuge I&M Project}
\usage{
create.dir(proj.name = NULL, dir.name = NULL, more.folders = NULL)
}
\arguments{
\item{proj.name}{The name of the project to be used as the name of the root project folder.}

\item{dir.name}{Location to create new root project folder. Default is the current working directory.}

\item{more.folders}{List or vector containing character strings specifying full directory paths for additional
folders to be included in project directory.}
}
\description{
A function to create a new file directory structure. The directory structure is based off the Alaska I&M biometric project directory template.
}
\examples{
\dontrun{
create.dir(proj.name = "bear_survey", dir.name = "./put_it_here", more.folders = NULL)}
}
