% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refugetools.R
\docType{package}
\name{refugetools}
\alias{refugetools}
\alias{refugetools-package}
\title{refugetools: A package that contains helpful functions for planning and implementing refuge I&M surveys.}
\description{
The refugetools package contains three functions:
create.dir, mdeditor, and dublin.
}
\section{Create.dir function}{

The create.dir function creates a project working directory.
}

\section{Mdeditor function}{

The mdeditor function opens the mdEditor application for creating and editing data and project metadata (json files).
}

\section{dublin function}{

 The dublin function creates a metadata csv file to document the contents of the project working directory.
}

